+++ packages/contracts/src/psas.ts
@@ -1,8 +1,12 @@
 import { initContract } from "@ts-rest/core";
 import { z } from "zod";
-import { PSASchema } from "./schemas/psas";
+import { PSASchema } from "@monkeytype/schemas/psas";
 
-import { CommonResponses, meta, responseWithData } from "./schemas/api";
+import {
+  CommonResponses,
+  meta,
+  responseWithData,
+} from "@monkeytype/schemas/api";
 export const GetPsaResponseSchema = responseWithData(z.array(PSASchema));
 export type GetPsaResponse = z.infer<typeof GetPsaResponseSchema>;
