+++ packages/contracts/src/quotes.ts
@@ -7,16 +7,16 @@ import {
   MonkeyResponseSchema,
   responseWithData,
   responseWithNullableData,
-} from "./schemas/api";
+} from "@monkeytype/schemas/api";
 import {
   ApproveQuoteSchema,
   QuoteIdSchema,
   QuoteRatingSchema,
   QuoteReportReasonSchema,
   QuoteSchema,
-} from "./schemas/quotes";
-import { IdSchema, NullableStringSchema } from "./schemas/util";
-import { LanguageSchema } from "./schemas/languages";
+} from "@monkeytype/schemas/quotes";
+import { IdSchema, NullableStringSchema } from "@monkeytype/schemas/util";
+import { LanguageSchema } from "@monkeytype/schemas/languages";
 
 export const GetQuotesResponseSchema = responseWithData(z.array(QuoteSchema));
 export type GetQuotesResponse = z.infer<typeof GetQuotesResponseSchema>;
