+++ packages/contracts/src/require-configuration/index.ts
@@ -1,4 +1,4 @@
-import { Configuration } from "../schemas/configuration";
+import { Configuration } from "@monkeytype/schemas/configuration";
 
 type BooleanPaths<T, P extends string = ""> = {
   [K in keyof T]: T[K] extends boolean
