+++ packages/schemas/.eslintrc.cjs
@@ -1,0 +1,5 @@
+/** @type {import("eslint").Linter.Config} */
+module.exports = {
+  root: true,
+  extends: ["@monkeytype/eslint-config"],
+};
