+++ packages/schemas/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@monkeytype/contracts",
+  "name": "@monkeytype/schemas",
   "private": true,
   "scripts": {
     "dev": "tsup-node --watch",
@@ -12,12 +12,10 @@
     "lint": "npm run oxlint && npm run eslint"
   },
   "peerDependencies": {
-    "@ts-rest/core": "3.52.1",
     "zod": "3.23.8"
   },
   "devDependencies": {
     "@monkeytype/eslint-config": "workspace:*",
-    "@monkeytype/schemas": "workspace:*",
     "@monkeytype/tsup-config": "workspace:*",
     "@monkeytype/typescript-config": "workspace:*",
     "chokidar": "3.6.0",
@@ -25,8 +23,7 @@
     "madge": "8.0.0",
     "oxlint": "1.7.0",
     "tsup": "8.4.0",
-    "typescript": "5.5.4",
-    "vitest": "2.1.9"
+    "typescript": "5.5.4"
   },
   "exports": {
     ".": {
