+++ packages/schemas/src/configuration.ts
