+++ packages/schemas/tsup.config.js
@@ -1,0 +1,3 @@
+import { extendConfig } from "@monkeytype/tsup-config";
+
+export default extendConfig();
