+++ packages/schemas/vitest.config.js
@@ -1,0 +1,11 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    globals: true,
+    environment: "node",
+    coverage: {
+      include: ["**/*.ts"],
+    },
+  },
+});
